+++ tests/general.rs
@@ -34,11 +34,13 @@ fn run_test<F: ObjectiveFunction + Clone + 'static>(
         let result = cmaes_state.run();
         let evals = cmaes_state.function_evals();
 
-        if !(result.overall_best.value < options.tol_fun) {
-            failures.push((result.reason, result.overall_best.value));
+        for reason in &result.reasons {
+            *reasons.entry(*reason).or_insert(0) += 1;
         }
 
-        *reasons.entry(result.reason).or_insert(0) += 1;
+        if !(result.overall_best.value < options.tol_fun) {
+            failures.push((result.reasons, result.overall_best.value));
+        }
 
         total_evals += evals;